Move the File-name ownership test to its own file
The structuredClone ownership test shares blob.test.ts with a pre-existing
RSS-threshold leak test that flakes on the debug build (its `isASAN`
threshold check keys on the binary being named `bun-asan`, but `bun bd`
produces `bun-debug`, so the strict non-ASAN threshold is applied under
ASAN's inflated RSS). Isolating the ownership test keeps it runnable
without that unrelated flake.
